partial converison to new interface
